Car Colliders w Controllers
I've added the colliders with visible meshes, for show. The front collider has a script which connects the car and the car user controller, so that when the front collider starts and continues a collision, it controls the car (reverse to the left). On exit of the collision, normal control resumes.