Clean Up for merge to Master
This commit has a driveable car with mesh colliders (as triggers) attached to it. It displays values in the "car" game object in the inspector tool that will be used in the neural network as inputs/outputs for nodes. The walls were also raised (at some point lol) so that the car can only stay on the track.