Automated Driving
Added InputsOutputs.cs file which controls the inputs (or switches, from the colliders) and the outputs of the car (h, v, and handbrak). The current implementation has a very basic if this then that structure. No switches on is just drive straight, and move the opposite direction if a collider is on.