Added Neural Network File
Added NeuralNetwork.cs file. this file contains an input layer (the collider switches) and an output layer (which drives the car user controls of h, v, and handbrake). Updated the collision if blocks so that it drives a little better (just for show. This doesn't display a "neural network" just yet")